Fix errors in expulsion calls
- repatriate functions return a DispatchResult so we can just let
any errors from them propagate
- proposal does not need to be in voting period to be dismissed
with prejudice. Move the status check to the else branch. We do check
the `voting_opened_on` in `can_dismiss_expulsion_proposal_with_prejudice`
but if it is 'None' it'll be set to the current block so the voting
period end check will always fail anyway.